RealEstate/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.5
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-crispy==1.8.1
django-filter==2.4.0
django-js-asset==1.2.2
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.12.2
idna==3.2
oauthlib==3.1.1
Pillow==8.3.2
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
text-unidecode==1.3
typing-extensions==3.10.0.2
urllib3==1.26.6
